Removed Unicode Character 'HORIZONTAL ELLIPSIS' (U+2026)
http://www.fileformat.info/info/unicode/char/2026/index.htm

+++ GetMp3.py
@@ -33,14 +33,18 @@ for eachA in aList:
     if 'href' in eachA.attrs:
         href = eachA.attrs['href']
         if str(href).endswith('.mp3'):
-            print('Downloading', href)
             filename = root_folder+os.path.sep+eachA.text
             filename = filename.replace(' – TamilWire.com', '')
             filename = filename.replace('\u2013', '-')
-            img = requests.get(href)
-            print('saving to', filename)
-            f = open(filename, 'wb')
-            f.write(img.content)
-            f.close()
+            filename = filename.replace('\u2026', '')
+            if not os.path.exists(filename):
+                print('Downloading', href)
+                img = requests.get(href)
+                print('saving to', filename)
+                f = open(filename, 'wb')
+                f.write(img.content)
+                f.close()
+            else:
+                print(filename, 'already downloaded')
 
 print('All mp3 files from', soup.title, 'downloaded successfully')
